requirements.txt
asgiref==3.4.1
black==21.9b0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Django==3.2.12
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-filter==21.1
django-jazzmin==2.4.9
django-sequences==2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
flake8==4.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pytz==2021.3
regex==2021.10.21
sqlparse==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.2
whitenoise==5.3.0
